Создаёт модульную структуру
